The first in a sequence of several updates stumbling to a solution to the problem that the message box re invalid monitor doesn't appear on top of the application.
The cause turned out to be that dialog boxes run on their own event queue,
so the message box was being shown before the applications main event queue
was running and the 2 sync'd up somehow, after which dialog boxes appear
over the main application.

Things that didn't work:
- Starting the message thread after main application show() and before
  exec().
- Connecting an initial event posted on the application queue
  to  a slot on the MsgboxThread to that opens processing.
- creating a SerialMsgBox  each time a message is to be shown,
  instead of initialing a SerialMsgBox and reusing it
- ...

The eventual solution, which feels like a hack, is just to sleep for
a fraction of a second when starting the MsgboxQueue.

Files are littered with the detritus of each of these attempts and
requires major cleanup.

Now for the first commit:

global_state.h:
- new member variable _msgBoxThread that records the single MsgBoxThred instance.