Ensure that MOC processes src/base/monitor.h by explicitly including it in list of source files for add_executable()
Because the source code is organized so that the dependencies among
directories form a directed graph, file monitor.h is located in src/base,
not src/main. To ensure that monitor.h is processed by the MOC, it is
explicitly listed in the SOURCES definition used by add_executable().

Previously, compilation was forced by the hack of having a symbolic
link from src/main/monitor.h -> src/base/monitor.h.